Support authentication and authorization using AWS Cognito
Authentication is supported via AWS user pools, and authorization
through JWT access tokens provided by Cognito.

The sign-in page is provided remotely by Cognito. The user is
authenticated using the AWS user pool, and when successful, Cognito
redirects to the site's configured callback URL, passing a response
containing an access token, which is in JWT form. The JWT can be
decrypted using the user pool's public key to find the embedded
info (user id, username, token expiry etc). The access token is set
as a cookie so that it can be verified (to determine whether it is
still valid) whenever a page requires authorization. Responses are
provided over HTTPS using OAuth.

We are using:
- flask_awscognito for the AWS Cognito integration (e.g. to go to the
AWS hosted UI for sign-in, and after sign-in redirection, validate
the provided JWT)
- flask_jwt_extended to manage the access token cookies and check
JWT validity for each page that requires authorization
- pyjwt to decode the user pool's public key to set the initial
configuration for flask_jwt_extended

All pages aside from the main index page now require authorization.
Whenever a page is loaded, we ask flask_jwt_extended to check for
a valid JWT token in the session cookies.

The auth features are only enabled if FLASK_ENV='production'.

Unit tests now set the ALCHEMY_MODE environment variable before the
first reference to the main Flask application, in order to set the
testing configuration.

Still to-do:
- Implement a sign-out mechanism. Currently the access tokens will
continue to be valid for one hour (the default) after issue.
- Refresh access tokens periodically so they don't expire while the
user is still browsing.