Remove summary classes in data_manager.py
Replace StudentPaperSummary, ClazzPaperSummary and CohortPaperSummary
with PaperScoreTally and MultiPaperScoreTally. The html macros now
refer to 'tally' member variables instead of '*_summary'.

This avoids the need to create StudentPaperSummary when we only
need the total scores per class/cohort. It also reduces the number
of database queries by simplifying the score queries.